completed the updation of rooms
