+++ views/owner/viewHotels.hbs
@@ -106,7 +106,7 @@
                                             <th>Is Blocked</th>
                                             <th>Is Available</th>
                                             <th>Edit</th>
-                                            <th>Add Rooms</th>
+                                            <th>Add Category</th>
                                         </tr>
                                     </thead>
                                     <tbody>
@@ -136,7 +136,7 @@
                                                         style="color: #2ea3a1;"></i></a>
                                             </td>
                                             <td>
-                                                <a href="/owner/hotel/addrooms?id={{this._id}}"
+                                                <a href="/owner/hotel/addcategory?id={{this._id}}"
                                                     class="btn btn-primary">Add</a>
                                             </td>
                                         </tr>
@@ -239,7 +239,7 @@
 
             const addRoomsCell = document.createElement('td');
             const addRoomsLink = document.createElement('a');
-            addRoomsLink.href = '/owner/hotel/addrooms?id=' + hotel._id;
+            addRoomsLink.href = '/owner/hotel/addcategory?id=' + hotel._id;
             addRoomsLink.className = 'btn btn-primary';
             addRoomsLink.textContent = 'Add';
             addRoomsCell.appendChild(addRoomsLink);
