+++ views/partials/adminDashboardSide.hbs
@@ -26,8 +26,8 @@
                 </a>
             </li>
             <li class="nav-item">
-                <a class="nav-link" href="/owner/rooms">
-                    <span>Rooms</span>
+                <a class="nav-link" href="/owner/category">
+                    <span>Category</span>
                 </a>
             </li>
             <li class="nav-item">
